# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "KRIAL API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
...
        "/catalog/categories/{category_code}": {
            "get": {
                "tags": [
                    "Категории"
                ],
                "summary": "Получить категорию по ее коду с преимуществами",
                "operationId": "78599cce9b3b284049b61aa4c9904c09",
                "parameters": [
                    {
                        "name": "category_code",
                        "in": "path",
                        "description": "Уникальный код категории",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "dg"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/catalog/categories/{category_code}/banners": {
            "get": {
                "tags": [
                    "Категории"
                ],
                "summary": "Получить маркетинговые баннеры по коду категории",
                "operationId": "1748095c28ff0b21534ca714992bebf4",
                "parameters": [
                    {
                        "name": "category_code",
                        "in": "path",
                        "description": "Уникальный код категории",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "dg"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
    }
}